fix receivable accounts

+++ GPA.Dtos/Invoice/InvoiceWithReceivableAccountsDto.cs
@@ -3,11 +3,11 @@
     public class InvoiceWithReceivableAccountsDto
     {
         public Guid InvoiceId { get; set; }
-        public int InvoiceStatus { get; set; }
-        public int SaleType { get; set; }
+        public byte InvoiceStatus { get; set; }
+        public byte SaleType { get; set; }
         public DateTime Date { get; set; }
         public string InvoiceNote { get; set; }
-        public int PaymentStatus { get; set; }
+        public byte PaymentStatus { get; set; }
         public decimal Payment { get; set; }
         public string ClientName { get; set; }
         public string ClientIdentification { get; set; }
